move stream handler into datasource package
