+++ pkg/datasource/stream.go
@@ -1,4 +1,4 @@
-package main
+package datasource
 
 import (
 	"context"
